src/main/java/com/lt/proxy/application/ProxyController.java
package com.lt.proxy.application;

import java.util.Map;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;
import com.lt.proxy.service.ProxyService;

@RestController
@RequestMapping("/proxy")
public class ProxyController {

    private final ProxyService proxyService;

    public ProxyController(ProxyService proxyService) {
        this.proxyService = proxyService;
    }

    @GetMapping(
        produces = MediaType.APPLICATION_JSON_VALUE
    )
    public ResponseEntity<Map<String, String>> getResponseHeaders(@RequestParam("proxy-url") String proxyUrl) {
        return ResponseEntity.ok(proxyService.getResponseHeaders(proxyUrl));
    }
}
